+++ src/components/Button/Button.tsx
@@ -3,7 +3,7 @@ import { ButtonProps } from './types';
 
 export const Button = ({ variant = 'contained', color = 'primary', size = 'lg', children, ...props }: ButtonProps) => {
   return (
-    <Styled.Button variant={variant} color={color} size={size} {...props}>
+    <Styled.Button $variant={variant} $color={color} $size={size} {...props}>
       {children}
     </Styled.Button>
   );